Optimised code of activity page
